Correcao de buscas feita
