+++ .env.example
@@ -1,0 +1,19 @@
+# BANCO LOCAL (DEVE SER CRIADO USANDO O COMPOSE DO REPOSITORIO DATABASE) NÃO PRECISA DE ALTERAÇÃO
+
+#DATABASE_HOST=localhost
+#DATABASE_USERNAME=root
+#DATABASE_PASSWORD=rootpassword
+#DATABASE_DB=app_db
+#DATABASE_PORT=3306
+
+# BANCO DE DESENVOLVIMENTO
+
+DATABASE_HOST=mysql-banco-dev-lp.alwaysdata.net
+
+# ADICIONE AS CREDENCIAIS
+DATABASE_USERNAME=
+DATABASE_PASSWORD=
+
+DATABASE_DB=banco-dev-lp_bortonedev
+DATABASE_PORT=3306
+
