+++ src/modules/blog/blog.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Get, Logger, Post, Query } from '@nestjs/common';
+import { Body, Controller, Logger, Post } from '@nestjs/common';
 import { BlogService } from './blog.service';
 import { BlogCreateDto } from './dto/blog-create.dto';
 
@@ -6,14 +6,6 @@ import { BlogCreateDto } from './dto/blog-create.dto';
 export class BlogController {
   private readonly logger = new Logger(BlogController.name);
   constructor(private readonly blogService: BlogService) {}
-
-  @Get()
-  listarPosts(@Query('termo') termo?: string) {
-    this.logger.log(
-      `Listando posts do blog com filtro: ${termo ?? 'sem filtro'}`,
-    );
-    return this.blogService.listarPosts(termo);
-  }
 
   @Post()
   criarPost(@Body() blogDto: BlogCreateDto) {
