+++ src/modules/busca/busca.service.ts
@@ -87,5 +87,84 @@ export class BuscaService {
       ).padStart(2, '0')}`,
       key: ano * 10000 + mes * 100 + dia,
     };
+  };
+
+  async listarBlog(termo?: string): Promise<{
+    itens: Blog[];
+    mensagem?: string;
+  }> {
+    const termoNormalizado = termo?.trim();
+
+    if (!termoNormalizado) {
+      const itens = await this.blogModel.findAll({ order: [['id', 'DESC']] });
+
+      return {
+        itens,
+        mensagem:
+          itens.length === 0 ? 'Nenhum item foi encontrado.' : undefined,
+      };
+    }
+
+    const filtros: Array<Record<string, unknown>> = [
+      { titulo: { [Op.like]: `%${termoNormalizado}%` } },
+      { conteudo: { [Op.like]: `%${termoNormalizado}%` } },
+    ];
+
+    const termoEhInteiroDecimal = /^(0|[1-9]\d*)$/.test(termoNormalizado);
+    if (termoEhInteiroDecimal) {
+      const termoComoNumero = parseInt(termoNormalizado, 10);
+      if (!Number.isNaN(termoComoNumero)) {
+        filtros.push({ id: termoComoNumero });
+      }
+    }
+
+    const itens = await this.blogModel.findAll({
+      where: { [Op.or]: filtros },
+      order: [['id', 'DESC']],
+    });
+
+    return {
+      itens,
+      mensagem: itens.length === 0 ? 'Nenhum item foi encontrado.' : undefined,
+    };
+  }
+
+  async listarCarrossel(termo?: string): Promise<{
+    itens: Banner[];
+    mensagem?: string;
+  }> {
+    const termoNormalizado = termo?.trim();
+
+    if (!termoNormalizado) {
+      const itens = await this.bannerModel.findAll({ order: [['id', 'DESC']] });
+
+      return {
+        itens,
+        mensagem:
+          itens.length === 0 ? 'Nenhum item foi encontrado.' : undefined,
+      };
+    }
+
+    const filtros: Array<Record<string, unknown>> = [
+      { descricao: { [Op.like]: `%${termoNormalizado}%` } },
+    ];
+
+    const termoEhInteiroDecimal = /^(0|[1-9]\d*)$/.test(termoNormalizado);
+    if (termoEhInteiroDecimal) {
+      const termoComoNumero = parseInt(termoNormalizado, 10);
+      if (!Number.isNaN(termoComoNumero)) {
+        filtros.push({ id: termoComoNumero });
+      }
+    }
+
+    const itens = await this.bannerModel.findAll({
+      where: { [Op.or]: filtros },
+      order: [['id', 'DESC']],
+    });
+
+    return {
+      itens,
+      mensagem: itens.length === 0 ? 'Nenhum item foi encontrado.' : undefined,
+    };
   }
 }
